applied a part of the large patch from Steve Buchi
Bug List:

- Tabbing between text boxes within the same movie clip.
- Fixed a bug with finding certain fonts in Windows.
- Fixed a bug where narrow text and a small font size would not be displayed correctly.
- Fixed a bug which occassionally caused a crach when flash variables are passed in during movie loading
- Added ability to hide text in a password text box.
- Fixed shift not being recognized in text box entry
- Fixed a bug in goto_frame: sometimes when rewinding frames, execute_state_reserve would remove the incorrect object.
- Now store and pass blend mode information to the renderer

+++ trunk/tu-testbed/gameswf/gameswf_sound.cpp
@@ -140,7 +140,7 @@ namespace gameswf
 
 	struct start_sound_tag : public execute_tag
 	{
-		Uint16	m_handler_id;
+		int	m_handler_id;
 		int	m_loop_count;
 		bool	m_stop_playback;
 		array<sound_envelope> m_envelopes;
